+++ src/index.ts
@@ -1,12 +1,16 @@
-export const version = '0.9.9';
+export const version = '0.9.10-SNAPSHOT';
 
+export { attr } from './attribute';
 export { BaseHTMLElement } from './c-base';
-export { all, append, AppendPosition, attr, className, closest, elem, first, frag, next, prev, style } from './dom';
+export { all, append, AppendPosition, closest, elem, first, frag, next, prev } from './dom';
 export { pull, puller, push, pusher } from './dx';
+export { escapeAttr, escapeHtml } from './escapes';
 export { addOnEvents, bindOnEvent, bindOnEvents, off, on, OnEvent, OnEventListener, OnListenerBySelector, trigger } from './event';
 export { addHubEvents, bindHubEvents, hub, Hub, HubBindings, unbindHubEvents } from './hub';
+export { className, style } from './style';
 export { customElement } from './ts-decorator-custom-element';
 export { onDoc, onEvent, onWin } from './ts-decorator-on-event';
 export { onHub } from './ts-decorator-on-hub';
 export { val } from './utils';
+