+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 
 ### v0.9.10 Upcoming
 
-- `.` demo - added `demo/` code sample see [DEMO (for live access)](https://demo.dom-native.org/core/index.html)
 - `+` event - added .shadowRoot binding when defined for `@onEvent`
+- `+` escape - added `escapeHtml(string)` and `escapeAttr(string)` (using the DOM element to do the escaping)
+- `.` demo - added `demo/` code sample see [DEMO (for live access)](https://demo.dom-native.org/core/index.html)
+- `.` demo - added `perf`
 
 
 ### v0.9.9 May 13th 2020